Fix warnings and organize interfaces
