+++ AzurePrOps/AzurePrOps.AzureConnection/Models/PullRequestComment.cs
@@ -1,0 +1,3 @@
+namespace AzurePrOps.AzureConnection.Models;
+
+public record PullRequestComment(string Author, string Content, DateTime PostedDate);
