+++ AzurePrOps/AzurePrOps.AzureConnection/Models/ReviewerInfo.cs
@@ -1,0 +1,3 @@
+namespace AzurePrOps.AzureConnection.Models;
+
+public record ReviewerInfo(string Id, string DisplayName, string Vote);
