+++ AzurePrOps/AzurePrOps.ReviewLogic/Models/FoldRegion.cs
@@ -1,0 +1,7 @@
+namespace AzurePrOps.ReviewLogic.Models;
+
+public class FoldRegion
+{
+    public int StartLine { get; set; }
+    public int EndLine { get; set; }
+}
