+++ AzurePrOps/AzurePrOps.ReviewLogic/Models/MetricData.cs
@@ -1,0 +1,7 @@
+namespace AzurePrOps.ReviewLogic.Models;
+
+public class MetricData
+{
+    public string Name { get; set; } = string.Empty;
+    public double Value { get; set; }
+}
