+++ AzurePrOps/AzurePrOps.ReviewLogic/Models/Notification.cs
@@ -1,0 +1,8 @@
+namespace AzurePrOps.ReviewLogic.Models;
+
+public class Notification
+{
+    public string Title { get; set; } = string.Empty;
+    public string Message { get; set; } = string.Empty;
+    public NotificationType Type { get; set; } = NotificationType.Info;
+}
