+++ AzurePrOps/AzurePrOps.ReviewLogic/Models/NotificationType.cs
@@ -1,0 +1,9 @@
+namespace AzurePrOps.ReviewLogic.Models;
+
+public enum NotificationType
+{
+    Info,
+    Warning,
+    Error,
+    Success
+}
