+++ AzurePrOps/AzurePrOps.ReviewLogic/Models/SearchResult.cs
@@ -1,0 +1,7 @@
+namespace AzurePrOps.ReviewLogic.Models;
+
+public class SearchResult
+{
+    public int LineNumber { get; set; }
+    public string Context { get; set; } = string.Empty;
+}
