+++ AzurePrOps/AzurePrOps.ReviewLogic/Models/SuggestionType.cs
@@ -1,0 +1,9 @@
+namespace AzurePrOps.ReviewLogic.Models;
+
+public enum SuggestionType
+{
+    Improvement,
+    BestPractice,
+    Performance,
+    Security
+}
