+++ AzurePrOps/AzurePrOps.ReviewLogic/Services/Interfaces/ICodeFoldingService.cs
